moved verificationTokens utils
