+++ src/pages/api/auth/manual-signin.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getUserFromEmail } from "../../../database/utils/users";
-import { isVerificationTokenWithEmail } from "../../../utils/verificationTokens";
+import { isVerificationTokenWithEmail } from "../../../database/utils/verificationTokens";
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
 import {setCookie} from 'nookies'
