+++ src/pages/api/auth/verifyemail/index.ts
@@ -1,5 +1,5 @@
 import {NextApiRequest, NextApiResponse} from 'next'
-import {getVerificationTokenWithToken, deleteVerificationToken} from '../../../../utils/verificationTokens'
+import {getVerificationTokenWithToken, deleteVerificationToken} from '../../../../database/utils/verificationTokens'
 import {createUserFromToken} from '../../../../database/utils/users'
 import jwt from 'jsonwebtoken'
 import {setCookie} from 'nookies'
