+++ src/pages/api/auth/verifyemail/resend.ts
@@ -1,7 +1,7 @@
 import {NextApiRequest, NextApiResponse} from 'next'
 import { sendToken } from '../../../../utils/emails'
 import {isUserWithEmail} from '../../../../database/utils/users'
-import {getVerificationTokenWithEmail} from '../../../../utils/verificationTokens'
+import {getVerificationTokenWithEmail} from '../../../../database/utils/verificationTokens'
 
 export default async function Resend(req:NextApiRequest, res:NextApiResponse) {
