+++ src/utils/auth.ts
@@ -3,7 +3,7 @@ import { getSession } from "next-auth/react";
 import { parseCookies } from "nookies";
 import { getUser, getUserFromEmail } from "../database/utils/users";
 import jwt from 'jsonwebtoken'
-import { isVerificationTokenWithEmail } from "./verificationTokens";
+import { isVerificationTokenWithEmail } from "../database/utils/verificationTokens";
 import { User } from "../database/interfaces";
 import axios from 'axios'
 import {signOut as nextAuthSignOut} from 'next-auth/react'
